refactor: move files related to Table
