+++ giraffe/src/components/SizedTable.tsx
@@ -13,7 +13,7 @@ import {LatestValueTransform} from './LatestValueTransform'
 import {newTableFromConfig} from '../utils/newTable'
 import {RawFluxDataTable} from './RawFluxDataTable'
 import {FluxTablesTransform} from './FluxTablesTransform'
-import {TableGraphLayer} from './TableGraphLayer'
+import {TableGraphLayer} from './Table'
 import {SimpleTableLayer} from './SimpleTable'
 
 import {usePlotEnv} from '../utils/usePlotEnv'
