+++ giraffe/src/components/Table/TableGraphs.scss
@@ -25,12 +25,12 @@ $table-light-graph--text-highlight: #383846;
   font-family: 'icomoon';
   font-style: normal;
   font-weight: normal;
-  src: url('../fonts/icomoon.eot') format('embedded-opentype');
-  src: url('../fonts/icomoon.eot') format('embedded-opentype'),
-    url('../fonts/icomoon.woff2') format('woff2'),
-    url('../fonts/icomoon.ttf') format('truetype'),
-    url('../fonts/icomoon.woff') format('woff'),
-    url('../fonts/icomoon.svg') format('svg');
+  src: url('../../fonts/icomoon.eot') format('embedded-opentype');
+  src: url('../../fonts/icomoon.eot') format('embedded-opentype'),
+    url('../../fonts/icomoon.woff2') format('woff2'),
+    url('../../fonts/icomoon.ttf') format('truetype'),
+    url('../../fonts/icomoon.woff') format('woff'),
+    url('../../fonts/icomoon.svg') format('svg');
 }
 
 .time-machine-tables {
