+++ giraffe/src/components/Table/TableGraphs.scss.d.ts
