+++ giraffe/src/components/Table/index.ts
@@ -1,0 +1,2 @@
+export {TableGraphLayer} from './TableGraphLayer'
+export {HoverTimeProvider} from './hoverTime'
